Moved DataLoader classes to separate file for ease of importing to other scripts.
